+++ gridstatus/ercot.py
@@ -17,7 +17,7 @@ from gridstatus.base import (
     Markets,
     NotSupported,
 )
-from gridstatus.decorators import ercot_update_dates, support_date_range
+from gridstatus.decorators import support_date_range
 from gridstatus.ercot_60d_utils import (
     process_dam_gen,
     process_dam_load,
@@ -78,6 +78,14 @@ RTM_PRICE_CORRECTIONS_RTID = 13045
 # https://www.ercot.com/mp/data-products/data-product-details?id=NP4-196-M
 DAM_PRICE_CORRECTIONS_RTID = 13044
 
+# LMPs by Electrical Bus
+# https://www.ercot.com/mp/data-products/data-product-details?id=NP6-787-CD
+LMPS_BY_ELECTRICAL_BUS_RTID = 11485
+
+# LMPs by Resource Nodes, Load Zones and Trading Hubs
+# https://www.ercot.com/mp/data-products/data-product-details?id=NP6-788-CD
+LMPS_BY_SETTLEMENT_POINT_RTID = 12300
+
 
 class ERCOTSevenDayLoadForecastReport(Enum):
     """
@@ -112,10 +120,6 @@ ACTUAL_SYSTEM_LOAD_BY_WEATHER_ZONE = 13101
 # Actual System Load by Forecast Zone
 # https://www.ercot.com/mp/data-products/data-product-details?id=NP6-346-CD
 ACTUAL_SYSTEM_LOAD_BY_FORECAST_ZONE = 14836
-
-# Historical DAM Clearing Prices for Capacity
-# https://www.ercot.com/mp/data-products/data-product-details?id=NP4-181-ER
-HISTORICAL_DAM_CLEARING_PRICES_FOR_CAPACITY_RTID = 13091
 
 # 60-Day DAM Disclosure Reports
 # https://www.ercot.com/mp/data-products/data-product-details?id=NP3-966-ER
@@ -177,6 +181,26 @@ class Document:
     publish_date: pd.Timestamp
     constructed_name: str
     friendly_name: str
+    friendly_name_timestamp: pd.Timestamp
+
+
+def parse_timestamp_from_string(date_str, time_str):
+    # Add a colon between hours, minutes, and seconds for pandas to parse
+    time_str = time_str[:2] + ":" + time_str[2:4]
+    if len(time_str) > 4:
+        time_str = time_str + ":" + time_str[4:6]
+
+    # Combine date and time strings
+    datetime_str = date_str + " " + time_str
+    # Convert to pandas timestampp
+    try:
+        timestamp = pd.to_datetime(datetime_str, format="%Y%m%d %H:%M:%S").tz_localize(
+            Ercot.default_timezone,
+            ambiguous=False,
+        )
+    except:  # noqa
+        timestamp = None
+    return timestamp
 
 
 class Ercot(ISOBase):
@@ -206,7 +230,6 @@ class Ercot(ISOBase):
     ACTUAL_LOADS_FORECAST_ZONES_URL_FORMAT = "https://www.ercot.com/content/cdr/html/{timestamp}_actual_loads_of_forecast_zones.html"  # noqa
     ACTUAL_LOADS_WEATHER_ZONES_URL_FORMAT = "https://www.ercot.com/content/cdr/html/{timestamp}_actual_loads_of_weather_zones.html"  # noqa
     LOAD_HISTORICAL_MAX_DAYS = 14
-    AS_PRICES_HISTORICAL_MAX_DAYS = 30
 
     def get_status(self, date, verbose=False):
         """Returns status of grid"""
@@ -517,7 +540,7 @@ class Ercot(ISOBase):
 
         doc = self._get_document(
             report_type_id=forecast_type.value,
-            published_before=None if date == "latest" else date,
+            published_before=date,
             constructed_name_contains="csv.zip",
             verbose=verbose,
         )
@@ -593,8 +616,9 @@ class Ercot(ISOBase):
 
         df["Delivery Interval"] = df["Delivery Interval"].astype("Int64")
         df = self.parse_doc(df, verbose=verbose)
-        df["Market"] = Markets.REAL_TIME_15_MIN.value
-        return self._finalize_spp_df(df, verbose=verbose)
+        return self._finalize_spp_df(
+            df, market=Markets.REAL_TIME_15_MIN, verbose=verbose
+        )
 
     def get_dam_spp(self, year, verbose=False):
         """Get Historical DAM Settlement Point Prices(SPPs)
@@ -619,8 +643,9 @@ class Ercot(ISOBase):
         df = pd.concat(all_sheets.values())
         # filter where DSTFlag == 10
         df = self.parse_doc(df, verbose=verbose)
-        df["Market"] = Markets.DAY_AHEAD_HOURLY.value
-        return self._finalize_spp_df(df, verbose=verbose)
+        return self._finalize_spp_df(
+            df, market=Markets.DAY_AHEAD_HOURLY, verbose=verbose
+        )
 
     def get_interconnection_queue(self, verbose=False):
         """
@@ -761,13 +786,89 @@ class Ercot(ISOBase):
 
         return queue
 
+    @support_date_range(frequency=None)
+    def get_lmp(
+        self,
+        date,
+        end=None,
+        market: str = None,
+        locations: list = "ALL",
+        location_type: str = "ALL",
+        verbose=False,
+    ):
+        """Get LMP data for ERCOT normally produced by SCED every five minutes
+
+        Can specify the location type to return "electrical bus" or "settlement point" data
+
+        """
+
+        if location_type.lower() == "Electrical Bus".lower():
+            report = LMPS_BY_ELECTRICAL_BUS_RTID
+        elif location_type.lower() == "Settlement Point".lower():
+            report = LMPS_BY_SETTLEMENT_POINT_RTID
+        import pdb
+
+        pdb.set_trace()
+        if end is None:
+            start = None
+            end = None
+            date = date
+        else:
+            start = date
+            end = end
+            date = None
+
+        docs = self._get_documents(
+            report_type_id=report,
+            date=date,
+            published_after=start,
+            published_before=end,
+            extension="csv",
+            verbose=verbose,
+        )
+
+        return self._handle_lmp(docs=docs, verbose=verbose)
+
+    def _handle_lmp(self, docs, verbose=False):
+        import pdb
+
+        pdb.set_trace()
+        df = self.read_docs(docs, verbose=verbose)
+
+        df = self._handle_sced_timestamp(df=df, verbose=verbose)
+
+        if "SettlementPoint" in df.columns:
+            df = self._handle_settlement_point_name_and_type(df, verbose=verbose)
+        elif "ElectricalBus" in df.columns:
+            # do same thing as settlement point but for electrical bus
+            df = df.rename(
+                columns={
+                    "ElectricalBus": "Location",
+                },
+            )
+            df["Location Type"] = "Electrical Bus"
+            # make Location string and location typoe category
+            df["Location"] = df["Location"].astype("string")
+            df["Location Type"] = df["Location Type"].astype("category")
+
+        df = df[
+            [
+                "SCED Timestamp",
+                "Location",
+                "Location Type",
+                "LMP",
+            ]
+        ]
+
+        return df
+
     @lmp_config(
         supports={
             Markets.REAL_TIME_15_MIN: ["latest", "today", "historical"],
             Markets.DAY_AHEAD_HOURLY: ["latest", "today", "historical"],
         },
     )
-    @support_date_range(frequency="DAY_START")
+    @support_date_range(frequency=None)
     def get_spp(
         self,
         date,
@@ -788,18 +889,62 @@ class Ercot(ISOBase):
             - ``Trading Hub``
             - ``Resource Node``
         """
+
+        publish_date = None
+        published_before = None
+        published_after = None
+        friendly_name_timestamp_before = None
+        friendly_name_timestamp_after = None
+
         if market == Markets.REAL_TIME_15_MIN:
-            df = self._get_spp_rtm15(
-                date,
-                verbose,
-            )
+            # no end, so assume requesting one day
+            if end is None:
+                # data is publish one day prior
+                friendly_name_timestamp_before = date.normalize() + pd.DateOffset(
+                    days=1
+                )
+                friendly_name_timestamp_after = date.normalize()
+
+            else:
+                friendly_name_timestamp_before = end
+                friendly_name_timestamp_after = date
+
+            report = SETTLEMENT_POINT_PRICES_AT_RESOURCE_NODES_HUBS_AND_LOAD_ZONES_RTID
         elif market == Markets.DAY_AHEAD_HOURLY:
-            df = self._get_spp_dam(date, verbose)
+            # no end, so assume requesting one day
+            if end is None:
+                # data is publish one day prior
+                publish_date = date.normalize() - pd.DateOffset(days=1)
+                published_before = None
+                published_after = None
+            else:
+                publish_date = None
+                published_before = end
+                published_after = date
+            report = DAM_SETTLEMENT_POINT_PRICES_RTID
+
+        docs = self._get_documents(
+            report_type_id=report,
+            date=publish_date,
+            published_before=published_before,
+            published_after=published_after,
+            friendly_name_timestamp_before=friendly_name_timestamp_before,
+            friendly_name_timestamp_after=friendly_name_timestamp_after,
+            constructed_name_contains="csv.zip",
+            verbose=verbose,
+        )
+
+        import pdb
+
+        pdb.set_trace()
+
+        df = self.read_docs(docs, verbose=verbose)
 
         return self._finalize_spp_df(
             df,
             locations=locations,
             location_type=location_type,
+            market=market,
             verbose=verbose,
         )
 
@@ -859,10 +1004,18 @@ class Ercot(ISOBase):
                 df.loc[is_load_zone_dc_tie_energy_weighted, "Location"] + "_EW"
             )
 
+        df["Location"] = df["Location"].astype("string")
+        df["Location Type"] = df["Location Type"].astype("category")
+
         return df
 
-    def _finalize_spp_df(self, df, locations=None, location_type=None, verbose=False):
+    def _finalize_spp_df(
+        self, df, market, locations=None, location_type=None, verbose=False
+    ):
         df = self._handle_settlement_point_name_and_type(df, verbose=verbose)
+
+        if market is not None:
+            df["Market"] = market.value
 
         df = df.rename(
             columns={
@@ -897,28 +1050,37 @@ class Ercot(ISOBase):
     def _get_spp_dam(
         self,
         date,
+        start,
+        end,
         verbose=False,
     ):
         """Get day-ahead hourly Market SPP data for ERCOT"""
-        if date == "latest":
+        if start == "latest":
             raise ValueError(
                 "DAM is released daily, so use date='today' instead",
             )
 
-        publish_date = utils._handle_date(date, self.default_timezone)
         # adjust for DAM since it's published a day ahead
-        publish_date = publish_date.normalize() - pd.DateOffset(days=1)
-        doc_info = self._get_document(
+
+        publish_date = None
+        published_before = None
+        published_after = None
+        if date:
+            publish_date = date.normalize() - pd.DateOffset(days=1)
+        elif start and end:
+            published_before = end
+            published_after = start
+
+        docs = self._get_documents(
             report_type_id=DAM_SETTLEMENT_POINT_PRICES_RTID,
             date=publish_date,
+            published_before=published_before,
+            published_after=published_after,
             constructed_name_contains="csv.zip",
             verbose=verbose,
         )
 
-        msg = f"Fetching {doc_info.url}"
-        log(msg, verbose)
-
-        df = self.read_doc(doc_info, verbose=verbose)
+        df = self.read_docs(docs, verbose=verbose)
 
         # fetch mapping
         # todo move to finalize?
@@ -950,15 +1112,7 @@ class Ercot(ISOBase):
         if len(docs) == 0:
             raise ValueError(f"Could not fetch SPP data for {date}")
 
-        all_dfs = []
-        for doc_info in tqdm.tqdm(docs, disable=not verbose):
-            doc_url = doc_info.url
-            msg = f"Fetching {doc_url}"
-            log(msg, verbose)
-            df = self.read_doc(doc_info, verbose=verbose)
-            all_dfs.append(df)
-
-        df = pd.concat(all_dfs).reset_index(drop=True)
+        df = self.read_docs(docs, verbose=verbose)
 
         df["Market"] = Markets.REAL_TIME_15_MIN.value
         return df
@@ -987,7 +1141,7 @@ class Ercot(ISOBase):
 
         return docs
 
-    @support_date_range(frequency="1Y", update_dates=ercot_update_dates)
+    @support_date_range(frequency="DAY_START")
     def get_as_prices(
         self,
         date,
@@ -1009,78 +1163,8 @@ class Ercot(ISOBase):
             pandas.DataFrame: A DataFrame with prices for "Non-Spinning Reserves", \
                 "Regulation Up", "Regulation Down", "Responsive Reserves", \
                 "ERCOT Contingency Reserve Service"
-
-        Source:
-            https://www.ercot.com/mp/data-products/data-product-details?id=NP4-181-ER
         """
 
-        # use to check if we need to pull daily files
-        if (
-            date.date()
-            >= (
-                pd.Timestamp.now(tz=self.default_timezone)
-                - pd.DateOffset(days=self.AS_PRICES_HISTORICAL_MAX_DAYS)
-            ).date()
-        ):
-            return self._get_as_prices_recent(date, end=end)
-        elif not end:
-            end = date
-
-        doc_info = self._get_document(
-            report_type_id=HISTORICAL_DAM_CLEARING_PRICES_FOR_CAPACITY_RTID,
-            constructed_name_contains=f"{date.year}.zip",
-            verbose=verbose,
-        )
-        doc = self.read_doc(doc_info, verbose=verbose)
-
-        doc = self._finalize_as_price_df(doc)
-
-        max_date = doc.Time.max().date()
-
-        df_list = [doc]
-
-        # if last df date is less than our specified end
-        # date, pull the remaining days. Will only be applicable
-        # if end date is within today - 3days
-        if max_date < end.date():
-            df_list.append(
-                self._get_as_prices_recent(
-                    start=max_date,
-                    end=end,
-                ),
-            )
-
-        # join, sort, filter and reset data index
-        data = pd.concat(df_list).sort_values(by="Interval Start")
-
-        data = (
-            data.loc[
-                (data.Time.dt.date >= date.date()) & (data.Time.dt.date <= end.date())
-            ]
-            .drop_duplicates(subset=["Interval Start"])
-            .reset_index(drop=True)
-        )
-
-        return data
-
-    @support_date_range("DAY_START")
-    def _get_as_prices_recent(self, date, verbose=False):
-        """Get ancillary service clearing prices in hourly intervals in Day
-            Ahead Market. This function can return the last 31 days
-            of ancillary pricing.
-
-        Arguments:
-            date (datetime.date, str): date of delivery for AS services
-
-            verbose (bool, optional): print verbose output. Defaults to False.
-
-        Returns:
-
-            pandas.DataFrame: A DataFrame with prices for "Non-Spinning Reserves", \
-                "Regulation Up", "Regulation Down", "Responsive Reserves", \
-                "ERCOT Contingency Reserve Service"
-
-        """
         # subtract one day since it's the day ahead market happens on the day
         # before for the delivery day
 
@@ -1098,12 +1182,12 @@ class Ercot(ISOBase):
 
         doc = self.read_doc(doc_info, verbose=verbose)
 
-        data = self._finalize_as_price_df(
+        df = self._finalize_as_price_df(
             doc,
             pivot=True,
         )
 
-        return data
+        return df
 
     @support_date_range("DAY_START")
     def get_60_day_sced_disclosure(self, date, end=None, process=False, verbose=False):
@@ -1504,7 +1588,7 @@ class Ercot(ISOBase):
         """
         doc = self._get_document(
             report_type_id=WIND_POWER_PRODUCTION_HOURLY_AVERAGED_ACTUAL_AND_FORECASTED_VALUES_RTID,
-            published_before=None if date == "latest" else date,
+            published_before=date,
             extension="csv",
             verbose=verbose,
         )
@@ -1534,7 +1618,7 @@ class Ercot(ISOBase):
 
         doc = self._get_document(
             report_type_id=SOLAR_POWER_PRODUCTION_HOURLY_AVERAGED_ACTUAL_AND_FORECASTED_VALUES_BY_GEOGRAPHICAL_REGION_RTID,
-            published_before=None if date == "latest" else date,
+            published_before=date,
             extension="csv",
             verbose=verbose,
         )
@@ -1582,7 +1666,7 @@ class Ercot(ISOBase):
         doc = self._get_document(
             report_type_id=HOURLY_RESOURCE_OUTAGE_CAPACITY_RTID,
             extension="csv",
-            published_before=None if date == "latest" else date,
+            published_before=date,
             verbose=verbose,
         )
 
@@ -1826,26 +1910,45 @@ class Ercot(ISOBase):
 
         return df
 
-    def _handle_sced_system_lambda(self, docs, verbose):
-        log("Reading SCED System Lambda files", verbose)
-
-        df = pd.concat(
-            [pd.read_csv(i.url, compression="zip") for i in tqdm.tqdm(docs)],
+    def _handle_sced_timestamp(self, df, verbose=False):
+        df = df.rename(
+            columns={
+                "SCEDTimeStamp": "SCED Timestamp",
+                "SCEDTimestamp": "SCED Timestamp",
+            },
         )
 
-        df["SCED Time Stamp"] = pd.to_datetime(df["SCEDTimeStamp"]).dt.tz_localize(
+        df["SCED Timestamp"] = pd.to_datetime(df["SCED Timestamp"]).dt.tz_localize(
             self.default_timezone,
             ambiguous=df["RepeatedHourFlag"] == "N",
         )
 
-        df["System Lambda"] = df["SystemLambda"].astype("float64")
+        df = df.drop("RepeatedHourFlag", axis=1)
 
-        df.drop(
-            ["SystemLambda", "SCEDTimeStamp", "RepeatedHourFlag"],
-            axis=1,
-            inplace=True,
+        return df
+
+    def _handle_sced_system_lambda(self, docs, verbose):
+        all_dfs = []
+        for doc in tqdm.tqdm(
+            docs, desc="Reading SCED System Lambda files", disable=not verbose
+        ):
+            log(f"Reading {doc.url}", verbose)
+            df = pd.read_csv(doc.url, compression="zip")
+            all_dfs.append(df)
+
+        df = pd.concat(all_dfs)
+
+        df = self._handle_sced_timestamp(df, verbose=verbose)
+
+        df["SystemLambda"] = df["SystemLambda"].astype("float64")
+
+        df = df.rename(
+            columns={
+                "SystemLambda": "System Lambda",
+            },
         )
-        df.sort_values("SCED Time Stamp", inplace=True)
+
+        df.sort_values("SCED Timestamp", inplace=True)
         return df
 
     @support_date_range("DAY_START")
@@ -1952,12 +2055,7 @@ class Ercot(ISOBase):
         return df
 
     def _handle_price_corrections(self, docs, verbose=False):
-        dfs = []
-        for d in tqdm.tqdm(docs, disable=not verbose):
-            df = self.read_doc(d, verbose=verbose)
-            dfs.append(df)
-
-        df = pd.concat(dfs)
+        df = self.read_docs(docs, verbose=verbose)
 
         df = self._handle_settlement_point_name_and_type(df)
 
@@ -1994,6 +2092,7 @@ class Ercot(ISOBase):
         self,
         report_type_id,
         date=None,
+        published_after=None,
         published_before=None,
         constructed_name_contains=None,
         extension=None,
@@ -2013,9 +2112,14 @@ class Ercot(ISOBase):
         if date == "latest":
             date = None
 
+        # no need to pass this
+        if published_before == "latest":
+            published_before = None
+
         documents = self._get_documents(
             report_type_id=report_type_id,
             date=date,
+            published_after=published_after,
             published_before=published_before,
             constructed_name_contains=constructed_name_contains,
             extension=extension,
@@ -2032,7 +2136,10 @@ class Ercot(ISOBase):
         self,
         report_type_id,
         date=None,
+        published_after=None,
         published_before=None,
+        friendly_name_timestamp_after=None,
+        friendly_name_timestamp_before=None,
         constructed_name_contains=None,
         extension=None,
         verbose=False,
@@ -2052,36 +2159,61 @@ class Ercot(ISOBase):
         for doc in docs:
             match = True
 
-            publish_date = pd.Timestamp(doc["Document"]["PublishDate"]).tz_convert(
-                self.default_timezone,
+            doc_url = f"https://www.ercot.com/misdownload/servlets/mirDownload?doclookupId={doc['Document']['DocID']}"  # noqa
+
+            parts = doc["Document"]["FriendlyName"].split("_")
+            try:
+                date_str = parts[1]
+                time_str = parts[2]
+                friendly_name_timestamp = parse_timestamp_from_string(
+                    date_str, time_str
+                )
+            except:
+                friendly_name_timestamp = None
+
+            doc_obj = Document(
+                url=doc_url,
+                publish_date=pd.Timestamp(doc["Document"]["PublishDate"]).tz_convert(
+                    self.default_timezone,
+                ),
+                constructed_name=doc["Document"]["ConstructedName"],
+                friendly_name=doc["Document"]["FriendlyName"],
+                friendly_name_timestamp=friendly_name_timestamp,
             )
 
+            if published_after:
+                match = match and doc_obj.publish_date > published_after
+
             if published_before:
-                match = match and publish_date <= published_before
+                match = match and doc_obj.publish_date <= published_before
 
-            if date:
-                match = match and publish_date.date() == date.date()
-
-            if extension:
-                match = match and doc["Document"]["FriendlyName"].endswith(extension)
-
-            if constructed_name_contains:
+            if friendly_name_timestamp_after and friendly_name_timestamp_before:
                 match = (
                     match
-                    and constructed_name_contains in doc["Document"]["ConstructedName"]
+                    and doc_obj.friendly_name_timestamp > friendly_name_timestamp_after
                 )
 
-            if match:
-                doc_id = doc["Document"]["DocID"]
-                url = f"https://www.ercot.com/misdownload/servlets/mirDownload?doclookupId={doc_id}"  # noqa
-                matches.append(
-                    Document(
-                        url=url,
-                        publish_date=publish_date,
-                        constructed_name=doc["Document"]["ConstructedName"],
-                        friendly_name=doc["Document"]["FriendlyName"],
-                    ),
+            if friendly_name_timestamp_before and friendly_name_timestamp_before:
+                match = (
+                    match
+                    and doc_obj.friendly_name_timestamp
+                    <= friendly_name_timestamp_before
                 )
+
+            if date and date != "latest":
+                match = match and doc_obj.publish_date.date() == date.date()
+
+            if extension:
+                match = match and doc_obj.friendly_name.endswith(extension)
+
+            if constructed_name_contains:
+                match = match and constructed_name_contains in doc_obj.constructed_name
+
+            if match:
+                matches.append(doc_obj)
+
+        if date == "latest":
+            return [max(matches, key=lambda x: x.publish_date)]
 
         return matches
 
@@ -2100,6 +2232,7 @@ class Ercot(ISOBase):
 
         doc_info = self._get_document(
             report_type_id=SETTLEMENT_POINTS_LIST_AND_ELECTRICAL_BUSES_MAPPING_RTID,
+            extension=None,
             verbose=verbose,
         )
         doc_url = doc_info.url
@@ -2120,6 +2253,13 @@ class Ercot(ISOBase):
         log(f"Reading {doc.url}", verbose)
         df = pd.read_csv(doc.url, compression="zip")
         return self.parse_doc(df, verbose=verbose)
+
+    def read_docs(self, docs, verbose=False):
+        dfs = []
+        for doc in tqdm.tqdm(docs, desc="Reading files", disable=not verbose):
+            log(f"Reading {doc.url}", verbose)
+            dfs.append(self.read_doc(doc, verbose=verbose))
+        return pd.concat(dfs).reset_index(drop=True)
 
     def parse_doc(self, doc, verbose=False):
         # files sometimes have different naming conventions